Added a new UI for Search; Made the homepage tabbed; added relavant classes/model/enum for searching
