+++ ExtendedMemory/Models/Location.cs
@@ -2,9 +2,9 @@
 {
     public class Location
     {
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
-        public string Street { get; set; }
+        //public double Latitude { get; set; }
+        //public double Longitude { get; set; }
+        //public string Street { get; set; }
         public string City { get; set; }
         public string State { get; set; }
         public string Country { get; set; }
