+++ ExtendedMemory/Models/SearchType.cs
@@ -1,0 +1,14 @@
+﻿namespace ExtendedMemory.Models
+{
+    public enum SearchType
+    {
+        Memory,
+        People,
+        Tag,
+        City,
+        State,
+        Country,
+        Date,
+        Time
+    }
+}
