+++ ExtendedMemory/Views/AboutPage.xaml.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-using Xamarin.Forms;
+﻿using Xamarin.Forms;
 
 namespace ExtendedMemory
 {
